+++ src/renderwindow.cpp
@@ -531,7 +531,7 @@ namespace Birb
 			/* Convert Vector2int points into Sint16 vectors */
 			Sint16* vx = new Sint16[points.size()];
 			Sint16* vy = new Sint16[points.size()];
-			for (int i = 0; i < (int)points.size(); i++)
+			for (size_t i = 0; i < points.size(); i++)
 			{
 				vx[i] = points[i].x;
 				vy[i] = points[i].y;
@@ -575,7 +575,7 @@ namespace Birb
 		bool DrawPolygon(const Color& color, const std::vector<Vector2f>& points)
 		{
 			Vector2int* intPoints = new Vector2int[points.size()];
-			for (int i = 0; i < (int)points.size(); i++)
+			for (size_t i = 0; i < points.size(); i++)
 			{
 				intPoints[i].x = std::round(points[i].x);
 				intPoints[i].y = std::round(points[i].y);
